refactor: mock 데이터 네이밍 케이스 변경 및 디렉토리 수정
